feat: AbstractController 도입
