+++ tomcat/src/main/java/com/techcourse/controller/HomeController.java
@@ -1,5 +1,6 @@
 package com.techcourse.controller;
 
+import org.apache.catalina.AbstractController;
 import org.apache.catalina.Controller;
 import org.apache.coyote.http11.request.Http11Request;
 import org.apache.coyote.http11.response.ContentType;
@@ -7,10 +8,10 @@ import org.apache.coyote.http11.response.Http11Response;
 import org.apache.coyote.http11.response.Http11ResponseBuilder;
 import org.apache.coyote.http11.response.StatusCode;
 
-public class HomeController implements Controller {
+public class HomeController extends AbstractController {
 
     @Override
-    public void service(Http11Request request, Http11Response response) throws Exception {
+    protected void doGet(Http11Request request, Http11Response response) throws Exception {
         Http11ResponseBuilder.build(response, StatusCode.OK, ContentType.TEXT_HTML_UTF8, "Hello world!");
     }
 }
