+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -2,39 +2,23 @@ package com.techcourse.controller;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
+import org.apache.catalina.AbstractController;
 import org.apache.catalina.Controller;
-import org.apache.coyote.http11.cookie.HttpCookie;
-import org.apache.coyote.http11.cookie.Session;
 import org.apache.coyote.http11.request.Http11Request;
 import org.apache.coyote.http11.request.HttpMethod;
 import org.apache.coyote.http11.response.ContentType;
 import org.apache.coyote.http11.response.Http11Response;
 import org.apache.coyote.http11.response.Http11ResponseBuilder;
 import org.apache.coyote.http11.response.StatusCode;
-import java.io.File;
-import java.net.URL;
-import java.nio.file.Files;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
 
-public class RegisterController implements Controller {
+public class RegisterController extends AbstractController {
+
     @Override
-    public void service(Http11Request request, Http11Response response) throws Exception {
-        HttpMethod method = request.getHttpMethod();
-        if (method.equals(HttpMethod.GET)) {
-            doGet(request, response);
-        }
-
-        if (method.equals(HttpMethod.POST)) {
-            doPost(request, response);
-        }
-    }
-
     protected void doGet(Http11Request request, Http11Response response) throws Exception {
         Http11ResponseBuilder.buildFile(response, StatusCode.OK, ContentType.TEXT_HTML_UTF8, "register.html");
     }
 
+    @Override
     protected void doPost(Http11Request request, Http11Response response) throws Exception {
         String account = request.getBodyValue("account");
         String email = request.getBodyValue("email");
@@ -42,6 +26,7 @@ public class RegisterController implements Controller {
 
         if (account == null || email == null || password == null) {
             Http11ResponseBuilder.buildFile(response, StatusCode.OK, ContentType.TEXT_HTML_UTF8, "register.html");
+            return;
         }
 
         User user = new User(account, email, password);
